tests/legacy_tests/database/requirements.txt
# Database testing requirements
# Inherits from users test requirements but focused on database testing

# Core testing framework
pytest>=7.4.0
pytest-cov>=4.1.0

# Database testing
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Date/time handling
python-dateutil>=2.8.0
